Conflicts:
	android/cli/commands/_build.js
	android/titanium/src/java/org/appcelerator/titanium/TiC.java
	iphone/Classes/TiUITableViewRowProxy.m
	iphone/cli/commands/_build.js

+++ iphone/Classes/TiUIButtonProxy.m
@@ -78,7 +78,7 @@
 
 -(CGFloat) verifyWidth:(CGFloat)suggestedWidth
 {
-	switch(styleCache)
+	switch((int)styleCache)
 	{
 		case UITitaniumNativeItemInfoLight:
 		case UITitaniumNativeItemInfoDark:
@@ -94,7 +94,7 @@
 
 -(CGFloat) verifyHeight:(CGFloat)suggestedHeight
 {
-	switch(styleCache)
+	switch((int)styleCache)
 	{
 		case UITitaniumNativeItemInfoLight:
 		case UITitaniumNativeItemInfoDark:
@@ -111,7 +111,7 @@
 
 -(UIViewAutoresizing) verifyAutoresizing:(UIViewAutoresizing)suggestedResizing
 {
-	switch (styleCache)
+	switch ((int)styleCache)
 	{
 		case UITitaniumNativeItemInfoLight:
 		case UITitaniumNativeItemInfoDark:
